Move overview card title dates
https://issues.redhat.com/browse/COST-1721
